Move persistence test helpers to its own file
